MDEV-18450 Slaves wait shutdown
The patches features an optional shutdown behavior to hold on
until after all connected slaves have been sent the last binlogged event.
The connected slave is one whose START SLAVE has been acknowledged and
that was not stopped since that though it could be technically
reconnecting in background.

The solution therefore disallows killing the dump thread until is
has found EOF of the latest binlog file.
It is up to the shutdown requester (DBA) to set up a sufficiently large
shutdown timeout value for shudown to wait patiently until
lagging behind slaves have been synchronized. On the other hand if a
specific slave needs exclusion from synchronization the DBA would have
to stop it manually which would terminate its dump thread.

`mysqladmin shutdown' is extended with a `--wait_for_slaves' option
which translates to `SHUTDOWN wait_for_slaves' sql query
to enable the feature on the client side.